backend/pyproject.toml
[tool.poetry]
name = "restaurant"
version = "0.1.0"
description = ""
authors = ["s13-19-m-python"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
pillow = "^10.2.0"
djangorestframework-simplejwt = "^5.3.1"
django-model-utils = "^4.4.0"
cloudinary = "^1.38.0"
six = "^1.16.0"
python-decouple = "^3.8"
django-cloudinary-storage = "^0.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.2.0"
pre-commit = "^3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
